BUGFIX Check if get_data_branch_dir_root_data exists

+++ src/scraper/abstract_doc/pipeline/AbstractDocPipelineCleanupMixin.py
@@ -20,16 +20,17 @@ class AbstractDocPipelineCleanupMixin:
 
         # cleanup get_data_branch_dir_root_data()
         # should only have get_dir_docs_for_cls()
-        for child_name in os.listdir(cls.get_data_branch_dir_root_data()):
-            child_path = os.path.join(
-                cls.get_data_branch_dir_root_data(), child_name
-            )
-            if child_name != cls.get_doc_class_label():
-                if os.path.isdir(child_path):
-                    shutil.rmtree(child_path)
-                else:
-                    os.remove(child_path)
-                log.warning(f"🧹 Deleted {child_path}")
+        if os.path.exists(cls.get_data_branch_dir_root_data()):
+            for child_name in os.listdir(cls.get_data_branch_dir_root_data()):
+                child_path = os.path.join(
+                    cls.get_data_branch_dir_root_data(), child_name
+                )
+                if child_name != cls.get_doc_class_label():
+                    if os.path.isdir(child_path):
+                        shutil.rmtree(child_path)
+                    else:
+                        os.remove(child_path)
+                    log.warning(f"🧹 Deleted {child_path}")
 
         # cleanup cls.get_dir_docs_for_cls()
         if os.path.exists(cls.get_dir_docs_for_cls()):
